Inline `@babel/highlight` in `@babel/code-frame`
