--- packages/babel-highlight/.npmignore
@@ -1,5 +1,0 @@
-src
-test
-*.log
-tsconfig.json
-tsconfig.tsbuildinfo
