--- packages/babel-highlight/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@babel/highlight",
-  "version": "7.25.9",
-  "description": "Syntax highlight JavaScript strings for output in terminals.",
-  "author": "The Babel Team (https://babel.dev/team)",
-  "homepage": "https://babel.dev/docs/en/next/babel-highlight",
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/babel/babel.git",
-    "directory": "packages/babel-highlight"
-  },
-  "main": "./lib/index.js",
-  "dependencies": {
-    "@babel/helper-validator-identifier": "workspace:^",
-    "chalk": "condition:BABEL_8_BREAKING ? : ^2.4.2",
-    "js-tokens": "condition:BABEL_8_BREAKING ? ^8.0.0 : ^4.0.0",
-    "picocolors": "^1.0.0"
-  },
-  "devDependencies": {
-    "strip-ansi": "^4.0.0"
-  },
-  "engines": {
-    "node": ">=6.9.0"
-  },
-  "conditions": {
-    "BABEL_8_BREAKING": [
-      {
-        "engines": {
-          "node": "^18.20.0 || ^20.17.0 || >=22.8.0"
-        }
-      },
-      {
-        "exports": null
-      }
-    ],
-    "USE_ESM": [
-      null,
-      {
-        "type": "commonjs"
-      }
-    ]
-  },
-  "exports": {
-    ".": {
-      "types": "./lib/index.d.ts",
-      "default": "./lib/index.js"
-    },
-    "./package.json": "./package.json"
-  },
-  "type": "module"
-}
