--- packages/babel-highlight/tsconfig.json
@@ -1,17 +1,0 @@
-/* This file is automatically generated by scripts/generators/tsconfig.js */
-{
-  "extends": [
-    "../../tsconfig.base.json",
-    "../../tsconfig.paths.json"
-  ],
-  "include": [
-    "../../packages/babel-highlight/src/**/*.ts",
-    "../../lib/globals.d.ts",
-    "../../scripts/repo-utils/*.d.ts"
-  ],
-  "references": [
-    {
-      "path": "../../packages/babel-helper-validator-identifier"
-    }
-  ]
-}
